requirements.txt
ansible
fedmsg
PyYAML
redis